[winstackers] Transition Ocata to End of Life
This transition the Ocata branch to End of Life. The last patch of the
branch will be tagged with ocata-eol tag. The ocata branch cannot be
used anymore and will be deleted if this patch merges.

This is needed as stable/ocata is not actively maintained in the recent
period and thus gate is broken due to job failures. By removing these
branches, infra resources will be freed up, too.

Please try to identify any zuul job, that is defined outside of the
repositories in this patch (for example in openstack-zuul-jobs, etc.)
and won't be used anymore if stable/ocata is deleted, and propose a job
removal patch for them.

Please +1 if the team is ready for us to proceed with this transition,
or -1 if there are still some activity on the branch and the team wants
to continue to maintain it.

Change-Id: I427d8f3607a8206c919d49ab3956598381c37511