Tag all tripleo-repos for X
Since we branched and tagged wallaby we have not done it on
master/Xena yet. This is causing a packaging problem for us as
described in the related-bug. We need the depends-on since that
is the commit we are tagging here for puppet-tripleo and validation
will not pass until that merges.

Related-Bug: 1931223
Depends-On: I9f3677fded010fcd659c154e5ec6a6a156e3fec3
Change-Id: If80932ebd102d4bcda70256f4cf82cb5267939da