Xena-2 Release for castellan
This is the Xena milestone-2 release for castellan.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 3.8.0..17cd833
1620394 barbican key manager: Add support for service user
2073f45 Changed minversion in tox to 3.18.0
0d0c2f2 Add missing font for PDF doc generation

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I4b7b3504426732f162c40502e8608102fa46df38