Xena-2 Release for metalsmith
This is the Xena milestone-2 release for metalsmith.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 1.4.2..407e505
407e505 Set image_disk_format from file extension for HTTP images
2fe1733 setup.cfg: Replace dashes with underscores
048ddb5 Stop testing the iscsi deploy interface
8cd0f73 Add Python3 xena unit tests
3114db1 Update master for stable/wallaby
621cd8c remove unicode from code

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I92adc1156365fe49f98df26377f5750a45cf322f