Xena-2 Release for octavia-lib
This is the Xena milestone-2 release for octavia-lib.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 2.3.1..cdbbacd
a3b01fd Fix loadbalancer_update argument typo
6a9cd92 Fix PDF docs build
53b4b8d Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I0106f83774821ee2e7c0b6d356981ecaf9bc3d2a