Xena-2 Release for openstacksdk
This is the Xena milestone-2 release for openstacksdk.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 0.57.0..9dc4a841
6c2a5ccb Move to OFTC
30600da9 Fix import order in accelerator and config unit tests
2e7b48ff tests: Rename 'TestProxyBase._verify2' to '_verify'
40cc84d5 Add IPsecSiteConnection resource and proxy layer functionality for it
6f985394 tests: Remove final use of 'TestProxyBase._verify'
2e726ef7 tests: Sanity check 'test_proxy_base'
5edbb15e tests: Remove dead code
38e9ca00 placement: Add support for resource classes

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I290e9aa7a1e9a42fc4c7c2f2466366e4c5a51fa7