Xena-2 Release for oslo.config
This is the Xena milestone-2 release for oslo.config.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 8.7.0..1a7bd66
e5fc313 config-generator yaml format doesn't work with i18n fields
b9468aa Ussuri+ is python3 only
ab8a8dd setup.cfg: Replace dashes with underscores

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ia291a5d496ae3fb71e41a6d7fc0b885d5a3c8aab