Xena-2 Release for oslo.middleware
This is the Xena milestone-2 release for oslo.middleware.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 4.2.0..43ab170
2c415ff Upgrade the pre-commit-hooks version
2ec633c setup.cfg: Replace dashes with underscores
b1ce931 Warning about /healthcheck
c7fabf4 Add Python3 xena unit tests
c1a30f3 Update master for stable/wallaby
3a0f7e4 Fix requirements issues
5c806d4 Fix requirements issues
2eb27aa remove unicode from code
ea0d1ab Use py3 as the default runtime for tox
a220331 ignore reno generated artifacts

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If8696416f6e757b8033bff79768dc1f5fe2ae96b