Xena-2 Release for python-ironic-inspector-client
This is the Xena milestone-2 release for python-ironic-inspector-client.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 4.5.0..82649e1
4046b74 [doc] Fix lldp_basic name
0687489 Expose more interface fields
b95e72c Add Python3 xena unit tests
db9bdd1 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I362996aedfd0f58e234df28b8d4bbde695904d56