Xena-2 Release for python-monascaclient
This is the Xena milestone-2 release for python-monascaclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 2.3.0..6d04502
6d04502 Fix Docker build
cf4d9cc Fix zuul publish docker image job
09709ed Add Python3 xena unit tests
d28ec21 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic704b8466e904c82e9de997886a668d4430090d2