Xena-2 Release for python-octaviaclient
This is the Xena milestone-2 release for python-octaviaclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by July 15.

$ git log --oneline --no-merges 2.3.0..6ef3eed
6ef3eed Fix some typos
84e927d Improve the client performance on large clouds
cf0c4be Support pagination for 'list' API calls
32a6c09 setup.cfg: Replace dashes with underscores
7927aec Add Python3 xena unit tests
7776264 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I0887820807ede79a0f3c3e7c4bff1b2cc5037c24