Release adjutant-ui for stable/ussuri
This release picks up new commits to adjutant-ui since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 0.5.0..400130e
400130e Now checks for code 202 when checking for success for other operations.
f2082dd Now checks for code 202 when checking for success
8a0d89c Comparison exception causes quota output to fail
073fbb9 Add .zuul.yaml file
47874bd Update TOX_CONSTRAINTS_FILE for stable/ussuri
c26b395 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I9860d80ad8752734c50c661be5bb4f6063e1cecb