Make stable/ussuri release for TripleO repos
Related puppet-tripleo metadata bump in depends-on is required
before validation will pass here as we are tagging that commit.
Some of these would warrant a bump to .feature rather than .patch
however we are locked to .patch on some of these for ussuri because
of an error made in an earlier (victoria) release.

Depends-On: Ia4aa0c87c1594f8e53c24a86d874998c561af931
Change-Id: Ied1c09bf9f6d22d9de7728ad4dfdff89e73f9ae0