Release barbican for stable/ussuri
This release picks up new commits to barbican since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 10.0.0..7cf483e2
7cf483e2 Allow multiple token labels for PKCS#11 driver
475074a4 Fix PKCS#11 reinitialization after failure
87822c14 Use system locks in pkcs11 library
35f13677 Drop lower-constraints job to unblock gate
48ea3222 Use barbican.conf in barbican-manage
1e2a1271 Use serial number or label for PKCS#11 tokens
c34591cd Rebase alembic migrations
5214aecc Update TOX_CONSTRAINTS_FILE for stable/ussuri
0c891b74 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I3199ca54b0dc88ae6f29bd5733f88dd52f6fa851