Release bifrost for stable/ussuri
This release picks up new commits to bifrost since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 8.1.1..2adb7bcc
2adb7bcc Use upper constraints when installing Bifrost requirements
b80a0b9c Change LANG in install-deps.sh to en_US.UTF-8
f1d0c916 Ensure that repositories are owned by the running users

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I605e15cf9781d4f903f6c884768a9c8053aefc2c