Release ceilometer for stable/ussuri
This release picks up new commits to ceilometer since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 14.0.0..f18675ae
ea7014ba [stable-only] Add font for PDF generation and fix lower constraints
baee38e9 Cap tenacity < 7.0.0
dafbcde0 Add logic for event_update handling and set transfer event as event_update
be8edb07 Run migration with NotFound exception as well
e92e4347 Run migration with NotFound exception as well
af90111c Remove [ and ] around parameter values in http url
96dcc2cb Fix PDF build
4185147d Trigger volume deletes only on volume deleted
fd41c23d Add volume.manage to metrics.
30f70243 Port the grenade job to Zuul v3
68e4686d Imported Translations from Zanata
27a1671c Update TOX_CONSTRAINTS_FILE for stable/ussuri
689215b4 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I7abdcbedb3c6d5f808a199195e6a76273604e19c