Release cloudkitty for stable/ussuri
This release picks up new commits to cloudkitty since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 12.1.0..f585666
f585666 Fix default admin_or_owner policy expression
7bfd8ec Drop lower-constraints job

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I4733f0e4d4360b61f42efb3c3e85de56216d2d85