Release cyborg for stable/ussuri
This release picks up new commits to cyborg since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..bb718c8
bb718c8 Resolve mysql conflict message
0b03b8b update API documentation for stable/ussuri
a35edb6 Fix hacking min version to 3.0.1
5b773fd Remove OPAE dependency in devstack
586ba30 Fix flake8 error from hacking <3.0.1
bdd2de0 Update TOX_CONSTRAINTS_FILE for stable/ussuri
a01613f Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I879c4bd6e05b7e44b75c72a4b0c437b6dca40ba9