Release designate for stable/ussuri
This release picks up new commits to designate since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 10.0.0..b967e9f7
b967e9f7 Update zones masters using pool target masters.
dc2426f7 Fix pdns4 devstack plugin
5a526c7a Drop lower-constraints test
06883ab7 pdns4 backend: check if zone exists before attempting delete
688987ab Fix lower-constraints test not constraining properly
9cf3b8ea fix database sync always get the wrong  value
a61a0c39 Update TOX_CONSTRAINTS_FILE for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I2c4005acc5b88b7055c80c987fa7df5f170b8218