Release glance for stable/ussuri
This release picks up new commits to glance since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 20.0.1..cc22468d
cc22468d Drop lower-constraints jobs
977ce4d7 Fix a failure to parse json file
3880298a Add a release note about import locking
5b681da1 Cleanup import status information after busting a lock
03a11dbb Add ImageLock to base flow checks
6c8f1e81 Functional test enhancement for lock busting
7ba0762d Add functional test for task status updating
bb573185 Handle atomic image properties separately
beece27e Move SynchronousAPIBase to a generalized location
173fdb19 Add FakeData generator test utility
28773650 Fix metadefs for compute-watchdog
055e5e79 Implement time-limited import locking
5998933a Add context.elevated() helper for getting admin privileges
ebeb31e6 Poll for final state on test_copy_image_revert_lifecycle()
541c0fd6 Fix non-deterministic copy_image_revert_lifecycle test
2d5e8439 Fix import failure status reporting when all_stores_must_succeed=True
931d39ec Functional reproducer for bug 1891352
3da696e0 Update task message during import
586ca78a Heartbeat the actual work of the task
825a0ec7 Add image_delete_property_atomic() helper
56d01f25 Add image_set_property_atomic() helper
663e027e Flesh out FakeImage for extra_properties
ec46cbca Add tests for _ImportToStore.execute()
ea34be42 Add testing for _CompleteTask in api_image_import
0895d191 Make test_copy_image_revert_lifecycle handle 409 on import retry
eb08fbae Make import task capable of running as admin on behalf of user
168fdb7f Add a functional test for non-owned image copying
3bf1af9b Refactor common auth token code in images test
bc826947 Add a test to replicate the owner-required behavior of copy-image

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I2708a7fb4d38f9f2528010fe9294ac66fa8b12bf