Release heat for stable/ussuri
This release picks up new commits to heat since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 14.1.0..b9a339597
b9a339597 Preserve order in list_concat_unique
e95a7fd21 Don't update status for replaced resource
fb879d1ee Net attr: Sort segments without name attribute first
49ef181bb Add SOURCE_IP_PORT to LB_ALGORITHM allowed values
e16fbbd13 Replace Fedora test image with F32
ec67396a2 As of nova microversion 2.57, personality is deprecated
936c6a50c Configure swap as before
8394d6c57 Supports 'virtio-forwarder' vnic type for port
18ab72f5d Use Fedora 31 for tests
18f2ed6b0 Fix OS::Heat::Delay resource type
4aafceb37 Lazily cache parsed value of list/json parameters

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ib97cc0f8592c36b03eb87b38455e659cd2de4e6d