Release ironic-inspector for stable/ussuri
This release picks up new commits to ironic-inspector since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 10.1.1..405576c
23494a4 Remove grenade jobs from old stable branches
f4fe4c7 Fix docs
5c40a5e Fix memcached host address.
c1a9478 Fix database migrations and disable the non-standalone job

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I377294c469ec992de93b1223d9b27aa16c1d1c00