Release ironic-prometheus-exporter for stable/ussuri
This release picks up new commits to ironic-prometheus-exporter since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.0.0..5744d42
5744d42 Remove lower-constraints job
1b15401 Fallback to `node_uuid` if`node_name` is not present
8dcc8ee Stop calling LOG.info on every function calls
2f328da Update TOX_CONSTRAINTS_FILE for stable/ussuri
125e361 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I9b668dd9a88a81162376e7b425fdd8bfa5e6d433