Release keystone for stable/ussuri
This release picks up new commits to keystone since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 17.0.0..2cf2912fd
8ab4eb27b Hide AccountLocked exception from end users
07d3a3d3f Retry update_user when sqlalchemy raises StaleDataErrors
6b739ffc3 Use app cred user ID in policy enforcement
d55c6c705 Drop lower-constraints job
f4819fe36 Support bytes type in generate_public_ID()
7ac089137 Delete system role assignments from system_assignment table
35c7406bf Implement more robust connection handling for asynchronous LDAP calls
c202bd506 Add vine to lower-constraints
d5870f69c Properly handle octet (byte) strings when converting LDAP responses
300e79e93 Fix lower-constraint for PyMySQL
1c37797c5 Port the grenade multinode job to Zuul v3
6b8e036ba Fix UserNotFound exception for expiring groups
1f0603598 Switch to new grenade job name

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I9f422804e0c005d705cea5f757d93a836fe48fe8