Release kuryr-kubernetes for stable/ussuri
This release picks up new commits to kuryr-kubernetes since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.0.0..50d6fd1e
a198e824 Fix duplicated sg rules on NP crd
81e53508 Correctly compare SG rules in NP code
01651084 Fix base docker images
4000025b Fix CI issues
43bf7ae1 Use latest RDO in container images
adf0d7cc Change registry for fetching coredns image.
3ee234c6 Move to quay.io with kuryr/demo docker image.
1862d232 Ignore errors when removing containers in DevStack
631364aa Handle None or {} labels in match_selector()
8cf37113 DevStack: Disable leader election for K8s services
b185cd2c Expand swap back to 8 GBs on gate VMs
9aa52355 Remove dragonflow
3b16d617 Fix CNI image built
ecf4aa6b Fix hacking min version to 3.0.1

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I8544dabe7b8d20675a5edbd5baa594fe3cea9d1d