Release kuryr-libnetwork for stable/ussuri
This release picks up new commits to kuryr-libnetwork since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.0.0..b2b1a36
2f05e8b Install tox before running tests
9b164c5 Handle subnet without gateway
080cd17 Update TOX_CONSTRAINTS_FILE for stable/ussuri
86a2816 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ia9634f65886acb46df8a07015ed63d02165c7719