Release kuryr for stable/ussuri
This release picks up new commits to kuryr since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.0.0..dbcca8a
2f021ec Remove run_as_root which fails without root_helper
6108958 Remove the unused coding style modules
90a458f Update TOX/UPPER_CONSTRAINTS_FILE for stable/ussuri
8c24aef Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ic516b58b1f73d350091cf0d7fa78f472abf300c8