Release magnum-ui for stable/ussuri
This release picks up new commits to magnum-ui since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.0.0..4a561a9
4a561a9 Fix cluster resize
1852c35 Fix npm jobs
f2c8f36 Imported Translations from Zanata
e1deec8 Imported Translations from Zanata
5bd5ced Imported Translations from Zanata
fd5e8fe Update TOX_CONSTRAINTS_FILE for stable/ussuri
bb0f5d8 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ic8c8691c1b91d6c086f8e627060283a9ccb15c26