Release manila-ui for stable/ussuri
This release picks up new commits to manila-ui since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.0..6054c99
21e3133 Fix share group creation with share types
a97f5e6 Fix parsing names in switched fields
117c3fd Drop lower constraints testing
0dbfc47 Remove nova networking
c29dd15 Drop legacy dsvm job & change zuulv3 integration job
24e0b14 Imported Translations from Zanata
331b1e4 Imported Translations from Zanata
5c47137 Update TOX_CONSTRAINTS_FILE for stable/ussuri
63a369f Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I0ce14f93b5c401384b5440c1a289a49e837152f4