Release metalsmith for stable/ussuri
This release picks up new commits to metalsmith since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.0.0..43bd9d4
43bd9d4 Remove lower-constraints job
4ee24fd Add the setuptools workaround
c658bce Use is instead of ambiguous pipe
3b32a3f CI: increase the swift loopback disk size for the CentOS job
d46002f Update TOX/UPPER_CONSTRAINTS_FILE for stable/ussuri
81eae8d Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I55aeeeaf355660fb8a541cb46b2a064f0bf0169e