Release monasca-agent for stable/ussuri
This release picks up new commits to monasca-agent since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.2..82839d9
82839d9 Fix Monasca-Setup Keystone detection
961cf3b Fix PY3 issue
859a46c Fix Docker builds
c9c0862 Fix small issue in dokcer plugin.
217c51a Remove incorrect plugins configuration for docker image.
6440156 Set 'libvirt_type' and 'libvirt_uri' options in the 'libvirt' plugin to be defined in conf.d/libvirt.yaml
1b8d426 Fix zuul publish docker image job

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ia58bdf2340ed523eee3fcaa0cd145bb88f0ece95