Release monasca-api for stable/ussuri
This release picks up new commits to monasca-api since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..36f18ae0
36f18ae0 Workaround for new pip 20.3 behavior
b43809ca Fix log-metric configuration
46f5d4a4 Fix docker publish image job
9ce0e0da Fix publish docker image job
ba45906e Align lower-constraints for new pip
ad13e892 Add option to choose Kafka client in Docker
8b146d20 Update gevent version to 1.5.0
5b145c2c Fix typo in http_check
080d1fea Update TOX_CONSTRAINTS_FILE for stable/ussuri
f310557a Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ibe5af762e63aea15b21e3899c5864d1542c81eca