Release monasca-notification for stable/ussuri
This release picks up new commits to monasca-notification since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.0.0..9dd0ea5
9dd0ea5 Fix Grafana URL in email notification
8f7a893 Fix zuul publish docker image job
05249ae Update lower-constraints
8fd51cd Fixed docker configuration template
3f9239e Add option to choose Kafka client in Docker
8151806 Remove deprecated YAML config
88df8d2 Set Docker publish jobs
caacf3b Update TOX_CONSTRAINTS_FILE for stable/ussuri
2fdf235 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I1d2ee0fef063d957a34f350d7e71ea47b6d1fc97