Release monasca-persister for stable/ussuri
This release picks up new commits to monasca-persister since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.0.0..2d4bd37
2d4bd37 Fix zuul publish docker image job
72c4097 Fix Docker build
9d64543 Add doc/requirements
dfe3b4d Update hacking for Python3
a6c50e5 Set legacy_kafka_client_enabled = False on default for Docker

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I548daa09f46e19f57bf2bb851e2e6fb90852622c