Release monasca-thresh for stable/ussuri
This release picks up new commits to monasca-thresh since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges train-em..cce79a5
7e8db2f Change storm config to eliminate storm warnings
48d0508 Fix zuul publish docker image job
bea100b Fix zuul publish docker image job for stable branche
909ccdd Enable usage of storm.yaml
e602fa7 Fix zuul publish docker image job
78d2f89 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I74063fd60ccd5503dcac6cacd114b4119c1f11f0