Release networking-hyperv for stable/ussuri
This release picks up new commits to networking-hyperv since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 8.0.0..e5b0ddb
e5b0ddb Monkey patch original current_thread _active

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I356edf876bc59531b59f85311fc71c439b6fae68