Release networking-sfc for stable/ussuri
This release picks up new commits to networking-sfc since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 10.0.0..ba57ca7
ba57ca7 Monkey patch original current_thread _active
90305b4 [stable] Drop lower-constraints job
3d4f698 Switch to stable/ussuri neutron-tempest-plugin jobs
0d4cb4f Update TOX_CONSTRAINTS_FILE for stable/ussuri
8f44ad9 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: If493da3a41208a7f47427178ac649b4623d344cd