Release octavia-dashboard for stable/ussuri
This release picks up new commits to octavia-dashboard since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.0.0..eac97f1
d0b91fd Drop lower-constraints job
b09a3d4 Disable Load Balancers panel when Octavia service is not deployed
7ab53ba Imported Translations from Zanata

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I51a3ca100dd41cc47a22f872ab359077b890100b