Release octavia for stable/ussuri
This release picks up new commits to octavia since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.2.0..e84d9478
e84d9478 Enable lo interface in the amphora-haproxy netns
de602782 Add amphora_id in store params for failover_amphora
45b31c18 Fix comment for the ca_certificates_file opt
4cfbf666 Optimize CountPoolChildrenForQuota task in amphorav2
d5dc2b26 Fix task_flow.max_workers with persistence in amphorav2
bae6808f Fix rsyslog configuration when disabling logs
8829f0fb Fix devstack cleanup when using amphorav2
19619ce8 Validate user access to vip_subnet_id when creating a LB
c7b4b875 Fix using subnets with host_routes in amphorav2 driver
170be0dd Make /healthcheck cache results
5b6d25ca Explicitely set nodeset to Bionic-based
46b9ce7c Cap hacking
7367311c Fix empty Batch Member Update to unlock objects
a8266a9e Fix weighted round-robin about UDP listener
7ded0e47 Fix LB failover for amphorav2: set security group
4b8de86c Fix incorrect ERROR status with IPv6 UDP members
dcce548e Fix provider driver quota handling


Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I7b8e581630d9e67f8bea0706e603ef569c28d64c