Release openstacksdk for stable/ussuri
This release picks up new commits to openstacksdk since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 0.46.0..1a7d8a32
1a7d8a32 Switch nodepool test to containers
e3da0210 Add id query parameter to sg rules
5028f7b7 Support roles 'name' in list_roles call
69d29cb3 Don't set list_type to dict for server groups.
d3020d94 Fix a bogus error in config loader when using several args with dashes
7149e50a OpenStack port decorator variables
3f45c9fa Don't error if clouds.yaml is not readable
14c1c394 Fix issues found by latest flake8
f4d27e4f Update docs to work with newer Sphinx
cc31e3a3 Update TOX/UPPER_CONSTRAINTS_FILE for stable/ussuri
8a03d530 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I28d3250c64c00049f74c598a163e7fb11c5ab5f1