Release os-brick for stable/ussuri
This release picks up new commits to os-brick since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.6..c336cb7
c336cb7 Avoid unhandled exceptions during connecting to iSCSI portals
5c031cb multipath/iscsi: remove devices from multipath monitoring
579e464 Drop lower-constraints job

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I7547ff40737f90aa3a43698e35644e6dfb5a3931