Release oslo.config for stable/ussuri
This release picks up new commits to oslo.config since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 8.0.4..63213f8
63213f8 remove lower-constraints.txt
be3558b config-generator yaml format doesn't work with i18n fields
f1ab258 Updating lower-constraints.txt

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I1bf1afac211cc95d1c0d0334cfcfcd1d6664f8f0