Release ovn-octavia-provider for stable/ussuri
This release picks up new commits to ovn-octavia-provider since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 0.1.2..4d20d32
4d20d32 Ensure that load balancer is added to logical switch
e749ac3 Add log for request_handler events
7d719a2 Remove the dependency on the "mock" package
de7c637 Check if no members in pool better in batch update
f4861f9 Don't try to store protocol=None in OVSDB
3204fcf Fix gate failure
e6c7e5a Fix leaked file descriptors by cleaning up objects
9aaf086 Retry status updates to Octavia
f98aaef Fix broken gate

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I54c56d187684a073a36fa476575558a246b6dcf8