Release ovsdbapp for stable/ussuri
This release picks up new commits to ovsdbapp since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.2.2..ae9e3fd
ae9e3fd Don't spam retries 100s of times a second
5d70023 Add an active wait in the "Backend.lookup"
5d18c40 Fix docs job
3b62043 Dropping lower constraints testing (stable Ussuri)

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I0d3ea4cf381cb44bfc7be2503e7e5b3231c1ecc4