Release placement for stable/ussuri
This release picks up new commits to placement since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.0..de030ef7
de030ef7 Fix Placement Doc
33591b83 Add DEBUG logs to help troubleshoot no allocation candidates
b08f3117 Switch to new grenade job name
6ee154e6 Update TOX_CONSTRAINTS_FILE for stable/ussuri
a8f981e3 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I640684aecc19f7e33d23dd276bb593b58619b59b