Release python-openstackclient for stable/ussuri
This release picks up new commits to python-openstackclient since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.2.1..c827e70f
0a8d8557 Fix reverted osc-lib interface change
595139ce Fix lower-constraints job
ec2f0f40 zuul: Stop testing against Tempest
e7ba80d2 Add API check for server_groups.list
441f116b Add API check for server_groups.create
b30843ce Bypass user and group verification in RemoveRole
59e0381b Add system role assignment tests for users and groups
5b25ea89 Make volume backup record commands available in v3
684cdc5b Replace assertItemsEqual with assertCountEqual

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ia3694c523ce207f4bb65552b28e4a7c6a1f4cfbd