Release python-zaqarclient for stable/ussuri
This release picks up new commits to python-zaqarclient since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.13.1..9750ac6
9750ac6 Drop python 2.7 support and testing

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ib36c5d1ab52c5a4276121bfa611ccb64c956b549