Release sahara-image-elements for stable/ussuri
This release picks up new commits to sahara-image-elements since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 12.0.0..80c09a5
80c09a5 Fix URL of Maven Central Repository
1d7279b Fix the buildimages jobs: call the ensure-tox role
ed9f7be Update TOX_CONSTRAINTS_FILE for stable/ussuri
59bb8ca Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ib378c1c111c1bbe8144afb28d196074a7cefa981