Release sahara-plugin-spark for stable/ussuri
This release picks up new commits to sahara-plugin-spark since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.0..46b812a
46b812a Fix URL of Maven Central Repository
beeef12 Update TOX_CONSTRAINTS_FILE for stable/ussuri
8018cf5 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I4ca84574f95567ba6d166462f9eff65f2cc9baa5