Release tacker for stable/ussuri
This release picks up new commits to tacker since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.0..5783b9ff
5783b9ff Fix : Register vim if use_barbican is disabled

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ic6886f8eb4e407bc247e7de1ab93bab9ca1d137b