Release trove for stable/ussuri
This release picks up new commits to trove since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 13.0.0..48e39a07
48e39a07 [docs]Fix wrong links in README
a5a5c5e0 Handle case where no networks are selected at create
7e492708 Set limit to -1 when getting Nova instances
d808241a [Image] Use sudo when removing image cache dir
83bf6ce1 Install pip3 for image build
f31ac746 [Backup] Ignore 404 error when deleting swift objects
695a63e2 Do not operate floating IP for relication cluster
c5d78f68 Fix building raw image
775abab5 Support image type for guest image building
8d9b938a Fix CI job
a006dc41 Update TOX_CONSTRAINTS_FILE for stable/ussuri
06755d64 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I18bdb098dc191361fd1148829c3e2ff3686ee063