Release watcher-dashboard for stable/ussuri
This release picks up new commits to watcher-dashboard since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.0..abcd7c0
abcd7c0 Fix broken devstack install
73dca39 Fix create audit fail from audit_template
0f00165 Update TOX_CONSTRAINTS_FILE for stable/ussuri
099ebde Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I46e72a13f0113776ade9e615ac78b34d1bc63006