Release watcher for stable/ussuri
This release picks up new commits to watcher since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..2e033d7b
2e033d7b Native Zuul v3 watcher-grenade job + some cleanup
0543ccfe Watcher API supports strategy name when creating audit template
cbb06e67 Check if scope is None
6e1be3cc Compatible with old scope format
043ebb71 Monkey patch original current_thread _active

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I31f72bbd062594e3f028ab08419f92dfb3c47483