Release zun for stable/ussuri
This release picks up new commits to zun since
the last release from stable/ussuri.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.0.0..5d41a395
5d41a395 api-ref: fix wrong naming in /execute
d5511fd8 Dep's should be restricted by upper-constraints
af2bd69f Pick up cafile from keystone_auth
f455bffc Fix init container status - docker
24f61108 Fix capsule's container status
261ae963 Update TOX_CONSTRAINTS_FILE for stable/ussuri
4a1c8ac6 Update .gitreview for stable/ussuri

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I322ddc18481d3145759399b283a66f30e01ca259