[heat] Propose to switch the release model
Intermediary-released deliverables that did release only once during
the last cycle, and have not done a release yet are good candidates
to switch to the cycle-with-rc model, which is much more suitable for
deliverables that are only released once per cycle.

These changes are not mandatory and this patch is a conveniences to
highlight this point and open the discussion.

PTLs and release liaisons please decide to:
- immediately release an intermediary release (and -1 the proposed change)
- confirm the release model change (+1 the proposed change)
- stay uncertain for this cycle of how many releases will be made, but
  acknowledge that they need to do a release before RC1 (-1 the proposed
  change)

Change-Id: I37dd7657632e12eaf0f0aadf5881b5dcbb154e22