[storlets] Transition Newton to Ocata to End of Life
This transition the Newton branch and the Ocata branch to End of Life.
The last patch of these branch will be tagged with *-eol tag. Theese
branchs cannot be used anymore and will be deleted if this patch
merges.

This is needed as these branches are not actively maintained in
the recent period and thus gate is broken due to job failures.
By removing these branches, infra resources will be freed up, too.

Change-Id: I93822aa5fd9a1dde6b8226b4bc48adbc8093c28b