Release final automaton for Xena
This proposed a release for automaton to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.3.0..0461d8e
0461d8e Moving to OFTC
df3f354 setup.cfg: Replace dashes with underscores
80bd6a5 Move flake8 as a pre-commit local target.
2353f50 Add Python3 xena unit tests
1703b2a Update master for stable/wallaby
0faae13 tox: Remove references to testr
283b8e2 Remove lower-constraints remnants

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I83550e91db0b8edb77cf03d0c178d7934a1ce634