Release final blazar-nova for Xena
This proposed a release for blazar-nova to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.2.0..965f178
965f178 Switch to stestr
668323a Sync requirements and lower constraints with nova
5f7bd6e Add Python3 xena unit tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I720f4f2502765f459d24ec8a158a5fa038a65871