Release final castellan for Xena
This proposed a release for castellan to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 3.9.0..4eb212b
4eb212b Moving IRC network reference to OFTC

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I63492f11606983f672011cafb03b21fe14fb0214