Release final cliff for Xena
This proposed a release for cliff to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 3.8.0..734bc0c
8fa916e Colourise and automatically page help output
9b66f40 Update unit test to satisfy python3.10+
392f3b2 Handle SIGPIPE exit gracefully
452fff3 Add conflict_handler parameter as attribut in Command class

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I044ea7a8cd9de0e6b467a8c1c751555cb4961569