Release final heat-translator for Xena
This proposed a release for heat-translator to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.4.0..0b8fba5
0b8fba5 [ussuri][goal] Update contributor documentation

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ieb5b7c3e8bcf8a92a07b10cb11f78729ea94deaf