Release final kuryr for Xena
This proposed a release for kuryr to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.3.0..764de32
0347de1 setup.cfg: Replace dashes with underscores
77fedd2 Add Python3 xena unit tests
0425000 Update master for stable/wallaby
8ef1757 remove unicode from code

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I58a3b8c17174c8ed9a822cfdf139b5b30c6a00d4