Release final metalsmith for Xena
This proposed a release for metalsmith to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 1.5.0..2a39acc6d5
d9c167d Use importlib instead of imp
e1e6cc3 Set glance limit for baremetal friendly images
419e384 Replace deprecated import of ABCs from collections
606c88b Changed minversion in tox to 3.18.0

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I5e2ada4bd525d9568d53eea58e1264adcc81dd84