Release final mistral-extra for Xena
This proposed a release for mistral-extra to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 11.0.0..acd70be
acd70be [community goal] Update contributor documentation
1b36297 Fix hacking version
f1d24ea Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I0c05c2e22a360cfe8be3bfa4cd986fb36b3ed1cd