Release final mistral-lib for Xena
This proposed a release for mistral-lib to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.4.0..44c33f2
44c33f2 [community goal] Update contributor documentation
2068c9c setup.cfg: Replace dashes with underscores
b49b6e4 Use py3 as the default runtime for tox
83b2d7e Dropping lower constraints testing
adb30bd Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ie8592c03aaf4b82f9159333b222cd78c23a430c1