Release final octavia-lib for Xena
This proposed a release for octavia-lib to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.4.0..ee07c4e
1ea6189 Add Python3 xena unit tests
c3fd4d1 Replace deprecated UPPER_CONSTRAINTS_FILE variable
aa03b89 setup.cfg: Replace dashes with underscores

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I6ffca1745b4d7fe54135119ff304a1abb8d6aa9d