Release final os-ken for Xena
This proposed a release for os-ken to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.0.0..0eb57b10
66d62049 setup.cfg: Replace dashes with underscores
042e32aa Replace deprecated UPPER_CONSTRAINTS_FILE variable

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I32dd0faa8e92d2432efcd9d12617a9e82dfa06fe