Release final os-vif for Xena
This proposed a release for os-vif to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.5.0..b837c1a
b837c1a add configurable per port bridges
b75af3a update os-vif ci to account for devstack default changes

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ica7db7d1bc036226584206e988509ac1706e3b6a