Release final os-win for Xena
This proposed a release for os-win to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 5.4.0..d571260
d571260 Use importlib instead of imp
50d0797 Use py3 as the default runtime for tox
530563b Add Python3 xena unit tests
a357177 Update master for stable/wallaby
80315f7 Use TOX_CONSTRAINTS_FILE

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ied2954434000d52b6eeab1a1af24a43bf789a392