Release final osc-lib for Xena
This proposed a release for osc-lib to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.4.1..415a6c7
4c3d3a0 Moving IRC network reference to OFTC
e3f58c1 Correct the tox option for skipping sdist generation

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I5b43120f291a14e0c37f8fcb9199947399f6e5f5