Release final oslo.context for Xena
This proposed a release for oslo.context to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 3.3.0..641a1e0
0fc0a0a Ussuri+ is python3 only and update python to python3

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I9631431d8a28ec8e243eb659b5026452efda2b4a