Release final oslo.messaging for Xena
This proposed a release for oslo.messaging to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 12.8.0..ee5eed86
bdbb6d62 Add Support For oslo.metrics
3e589edd Changed minversion in tox to 3.18.0
c1f9fbea Upgrade the pre-commit-hooks version
daca2549 setup.cfg: Replace dashes with underscores
39826f06 Remove the oslo_utils.fnmatch

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If894ff6d4cddcc7aec3ef9254113469e2c92b099