Release final oslo.metrics for Xena
This proposed a release for oslo.metrics to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 0.2.1..43eee50
29ded7b Enable Test Coverage and Add New Test
0a5cb7a Fix Heading Characters
86e1bb1 Enable bandit
d33cf58 Add Devstack Installation Guide
f03eb56 Use py3 as the default runtime for tox
3afeae8 Dropping lower constraints testing
e8868ad Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I0cb63e6f02301a88ddcbc136abb11f1830f811b7