Release final oslo.policy for Xena
This proposed a release for oslo.policy to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 3.8.1..c7fd9f4
c7fd9f4 Fix a typo in the document
fa5f76e Changed minversion in tox to 3.18.0

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I2d54646ddc5659b9a89c1adb6bf635f50871d9e4