Release final oslo.privsep for Xena
This proposed a release for oslo.privsep to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.6.0..9aac44a
fee00eb Changed minversion in tox to 3.18.0

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Id7ebd54c7d585550fff0e3e78a4ba3b9449dcd9c