Release final oslo.serialization for Xena
This proposed a release for oslo.serialization to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 4.1.0..2b94a4f
2b94a4f setup.cfg: Replace dashes with underscores
c587b7e Move flake8 as a pre-commit local target.
02a5e4d Add Python3 xena unit tests
2b41ddc Update master for stable/wallaby
a1f90ab Remove lower-constraints remnants
967a36f Drop use of deprecated collections classes
d20434c Use TOX_CONSTRAINTS_FILE

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I946c734c8958b641e211453289d7bb3f26bb9700