Release final oslo.utils for Xena
This proposed a release for oslo.utils to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 4.9.2..2c74bb9
2c74bb9 Modify UUID sentinel to support keystone-like UUIDs

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ife237e0d09beb2040af5934965100123258db223