Release final oslo.versionedobjects for Xena
This proposed a release for oslo.versionedobjects to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.4.0..a85275f
07fd2f2 Replace deprecated inspect.getargspec
63cdc14 Do not patch inspect.argspec directly
59c0aad Changed minversion in tox to 3.18.0
a1ecfaa Move flake8 as a pre-commit local target.
4fa96d6 Add Python3 xena unit tests
0a3da61 Update master for stable/wallaby
117d28e Dropping lower constraints references
e8062fc Dropping lower constraints testing
bdce5a0 Use TOX_CONSTRAINTS_FILE
7ce947c Use py3 as the default runtime for tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I27a4ec0615c6ca7291b25ad0aa9663503404c4c0