Release final ovsdbapp for Xena
This proposed a release for ovsdbapp to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 1.11.0..35c8eae
c8cb43e Improve "OvsdbIdl.from_server" method
f02b1c2 Make exceptions in the event logs easier to read
ee39ae0 setup.cfg: Replace dashes with underscores

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Idbeb9d09c345766b14e364dcfbaeca78c7ccec21