Release final stevedore for Xena
This proposed a release for stevedore to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 3.3.0..8846a3f
cba4821 setup.cfg: Replace dashes with underscores
ba1268d Fix formatting of release list
74d6d58 Remove lower-constraints remnants
5383853 Move flake8 as a pre-commit local target.
518948a Add Python3 xena unit tests
107308a Update master for stable/wallaby
90b5daa Dropping lower constraints testing
8b6b861 Use TOX_CONSTRAINTS_FILE

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Iff464b1b78839901ca1700cc0016ab558f5000b1