Release final sushy for Xena
This proposed a release for sushy to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 3.11.0..6669a29
6669a29 Add lower-constraints job to current development branch

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ida39290a960374fd2c726a11fb4f4770f93562b3