Release final tosca-parser for Xena
This proposed a release for tosca-parser to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

$ git log --oneline --no-merges 2.4.0..0014362
e7b9386 Fix old link in Installation guide
405ec35 Replace assertItemsEqual with assertCountEqual
92a3e69 import "collections.abc" explicitly
0501264 [ussuri][goal] Update contributor documentation
ca3a2a8 Fix parent types if relationships

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I22768e499ef808867ca097bd7e189a90d66112ef