Release final glance_store for Xena
This proposed a release for glance_store to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 23.

Change-Id: Ib13cbe6e35c83d89216ad24958c4004affba13cb