Release final oslo.messaging for Xena
This proposed a release for oslo.messaging to get any final merged changes
out prior to the lib freeze for Xena going into effect. We previously
announced a wrong date for the final lib series, this is a convenience
patch to allow to release latest changes before freezing the libs.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 30.

$ git log --oneline --no-merges 12.9.0..f9de265f
f9de265f amqp1: Do not reuse _socket_connection on reconnect
01f5b378 amqp1: re-organize TestFailover to be reused by TestSSL
fc49e04e Revert "Disable AMQP 1.0 SSL unit tests"

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic4256e444489b060b31fa5b7f3542a8044ed6958