Release final oslo.privsep for Xena
This proposed a release for oslo.privsep to get any final merged changes
out prior to the lib freeze for Xena going into effect. We previously
announced a wrong date for the final lib series, this is a convenience
patch to allow to release latest changes before freezing the libs.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 30.

$ git log --oneline --no-merges 2.6.1..a3256ac
4f14506 Add except path with exception debug to send_recv

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Iff10cee7c7fef06b855f9c536b47aaacd1b4646b