Release final sushy for Xena
This proposed a release for sushy to get any final merged changes
out prior to the lib freeze for Xena going into effect. We previously
announced a wrong date for the final lib series, this is a convenience
patch to allow to release latest changes before freezing the libs.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, August 30.

$ git log --oneline --no-merges 3.11.1..e8c29cc
e8c29cc Change default value: insert_media write_protected

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I804ca2a491b65f82bd6d6be61fb69e486ee1b0b0