Release final osc-placement for Xena
This proposed a release for osc-placement to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 3.0.1..HEAD
5e38a59 Add support for microversion 1.38 consumer types
8b3ed3c Mark microversion 1.37 supported

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Id3d28c9b49ccc5e08d36ce157d7acda5ae799be8