Release final python-blazarclient for Xena
This proposed a release for python-blazarclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 3.2.0..d2eb2f6
f2277b5 Add openstackclient support
27c426a Migrate from testr to stestr
c9114b7 Default client service type to reservation
1cf4d73 Add Python3 xena unit tests
c3fef3d Update master for stable/wallaby
a2d8abf Change dashes to underscore in setup.cfg

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I1aeb8ca05241dfe4c523f6f0d0bc1e3216eaf530