Release final python-cloudkittyclient for Xena
This proposed a release for python-cloudkittyclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 4.2.0..647f561
647f561 Replace deprecated import of ABCs from collections
61ef2c3 Fix creation of hashmap mapping with a zero cost
eb7f5e4 Changed minversion in tox to 3.18.0
0608c05 docs: Update Freenode to OFTC
754e06c setup.cfg: Replace dashes with underscores
d12e5a8 [ussuri][goal] Update contributor documentation
d2c2323 Fix PDF docs build
67fae0f Use py3 as the default runtime for tox
73668be Add release notes job template
dc606b1 Add Python3 xena unit tests
5141971 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ie80ba4cff4868485bee796bb81586a2702953c60