Release final python-cyborgclient for Xena
This proposed a release for python-cyborgclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 1.3.0..d051769
28f9622 Add test case for show arq
686ebac Fix the bug of making device_profile_group_id parameter is optional
ec2a079 Replace deprecated inspect.getargspec
9d5bdc5 Hide the traceback when deleting a non-existent ARQ
83c099d Add checks for showing dp with name
72272fe [ussuri][goal] Update contributor documentation
b5f9486 setup.cfg: Replace dashes with underscores
1d6d8e2 Add Python3 xena unit tests
1c0a3b4 Update master for stable/wallaby
2a674e6 bug fix: tox wrongly install client from repo path.
6c93f13 Add doc/requirements
8a9a48a update error url
18610ce Revert "remove py37"

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I9d7c76f1889c1b1c639ca6a94ad1c3af47e879e8