Release final python-glanceclient for Xena
This proposed a release for python-glanceclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 3.4.0..74a55bd
1eb0bbb Fix undesirable raw Python error
cb084f5 Add member-get command
158d5f4 Clean up extra spaces
b443996 setup.cfg: Replace dashes with underscores

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I41a47fac39a275db84a1d97d333013c89777a935