Release final python-heatclient for Xena
This proposed a release for python-heatclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 2.3.0..b438314
b438314 Replace oslo_utils.fnmatch with fnmatch
eff0d18 fix lower constraint job
cc2014f [ussuri][goal] Update contributor documentation
490e6b6 Add Python3 xena unit tests
3a7b038 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ib7121363276d52bcac53f2f2699767f4b52da3f6