Release final python-mistralclient for Xena
This proposed a release for python-mistralclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 4.2.0..608ec45
ab54cb9 Allow enforcing usage of "raw" definitions
ffa67c1 [community goal] Update contributor documentation
18fff74 Remove tripleo job from mistralclient
dba9c08 Use py3 as the default runtime for tox
87e3a7b Add Python3 xena unit tests
f1e217a Update master for stable/wallaby
63e20bf Drop lower-constraints job

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I08f04104c426134ae678c2aaca89f1109cebf130