Release final python-neutronclient for Xena
This proposed a release for python-neutronclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 7.5.0..HEAD
f83108d Use yaml.safe_load instead of yaml.load
2f047b1 Set ML2/OVS backend explicitly for functional job
23fb666 tests: change safe_hasattr to hasattr
a1ebfaa Ussuri+ is python3 only and update python to python3
d6c211c Changed minversion in tox to 3.18.0

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Iaded0070ca9769d4428eecf285ae975a6c2c1a6d