Release final python-novaclient for Xena
This proposed a release for python-novaclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 17.5.0..8066f8c7
8066f8c7 Microversion 2.90 - Configurable hostnames
01c7a3aa tests: Add missing 'nova update' unit tests
4a5bdde3 Add pre-commit
cfa172c4 Microversion 2.89 - os-volume_attachments
bff8d413 Use importlib instead of imp
273c4157 Use Block Storage API v3 instead of API v2

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If850205d0334205ca5c5e79c8f992b6bf386ed64