Release final python-tackerclient for Xena
This proposed a release for python-tackerclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 1.7.0..98789f8
98789f8 Multi version API support
6325127 Remove use of testtools.helpers.safe_hasattr
e8280f5 Remove broken link from Python-TackerClient guide
94e28fc Correct YAML load warning
c20ef24 vnfConfigurableProperties missing in VNF show
4c59029 Update min version of tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I82ebacef170592867182de6ecf7e8176447cdaa6