Release final python-troveclient for Xena
This proposed a release for python-troveclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 7.1.0..64f3f62
64f3f62 Fix detach configuration from instance
5ce2e99 Changed minversion in tox to 3.18.0

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic98587141f737b65a2bd57be50b08a1c8b6bdd44