Release final python-vitrageclient for Xena
This proposed a release for python-vitrageclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 4.3.0..bc829d3
bc829d3 setup.cfg: Replace dashes with underscores
096b484 Follow-up patch for fixing coverage issue
77bcc1b Use py3 as the default runtime for tox
63610f0 Migrate from testr to stestr
e70ec02 Add Python3 xena unit tests
3611f26 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I8740e7dce7ba90565b92bd8b781674cfdb317314