Release final python-watcherclient for Xena
This proposed a release for python-watcherclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 3.2.0..dc9b5cb
cf18f56 Changed minversion in tox to 3.18.0
f2f882e Switch testing to Xena testing runtime
add4254 Replace deprecated UPPER_CONSTRAINTS_FILE variable
6a4937d setup.cfg: Replace dashes with underscores
d0f4199 Use py3 as the default runtime for tox
52a3fd0 Drop lower constraints testing
c97e16f remove unicode from code

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I43742e4df4717707df42d87111b7ad3f7314e1ec