Release final python-zunclient for Xena
This proposed a release for python-zunclient to get any final merged changes
out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 02.

$ git log --oneline --no-merges 4.2.0..75e228d
7ca6a6b Uncap PrettyTable
6e7e579 setup.cfg: Replace dashes with underscores
6b6c09e Use py3 as the default runtime for tox
dc8db03 Add Python3 xena unit tests
49d0e41 Update master for stable/wallaby
bbc32db remove unicode from code

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ibc2e2c947122a4914e75ad8e078c65cdce0daec5