[neutron] Create for all client and non-client lib xena branch
This creates the stable/xena branch for all client
and non-client libraries now that we are past the
lib freeze on September 02.

If a team wants to wait for a specific patch to make it to the library,
someone from the team can -1 the patch to have it held,or update that
patch with a different commit SHA.

Patches with no response will be merged by Friday, September 10.

Change-Id: I3d77d00f0ea8635c8b6983655a7a269a247553bd