Proposing Xena RC1 for adjutant
This proposed a RC1 release for adjutant to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 2.0.0..7aa1fc6
7aa1fc6 Change default volume API to v3
d9045e4 Update formatting because of black version

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ib976f07ca7331faad9afdce83a0395bf21e78c3e