Proposing Xena RC1 for ansible-role-atos-hsm
This proposed a RC1 release for ansible-role-atos-hsm to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 1.0.0..c9f757d
c9f757d setup.cfg: Replace dashes with underscores

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ia13983466735b7f273865db36b8354c6842278a5