Proposing Xena RC1 for ansible-role-lunasa-hsm
This proposed a RC1 release for ansible-role-lunasa-hsm to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 1.1.0..26da379
26da379 Remove use of ansible_fqdn

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Iff52c571c3c574e1abdd8f862ec396c222229d37