Proposing Xena RC1 for blazar-dashboard
This proposed a RC1 release for blazar-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 5.0.0..5876f70
5876f70 Change minversion of tox to 3.18.0
46ac7b8 Fix PDF docs build
c337097 Add Python3 xena unit tests
2de6bc0 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ifa22b4eda4a858f0f82a884297b6481b3569657f