Proposing Xena RC1 for blazar
This proposed a RC1 release for blazar to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 7.0.0..233e1c1
339e913 Handle AggregateNotFound when deleting aggregate
21c0152 Add openstackclient command examples in CLI pages
63d9417 Remove six again
cfb561e docs: Update Freenode to OFTC
c9a4bbd Change minversion of tox to 3.18.0
4c1925e Add Python3 xena unit tests
d7b82a0 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I1c6f212d5d234c1f6b60ada71e1b00ccfb56e71b