Proposing Xena RC1 for cloudkitty-dashboard
This proposed a RC1 release for cloudkitty-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 12.0.0..d3fdbcd
d3fdbcd Replace deprecated import of ABCs from collections
2a0ec47 Moving IRC network reference to OFTC
53162a1 docs: Update Freenode to OFTC
01ff88b Changed minversion in tox to 3.18.0
3b02d94 [ussuri][goal] Update contributor documentation
57b185a setup.cfg: Replace dashes with underscores
19ead1f Respect endpoint type specified in Horizon
a92f956 Add Python3 xena unit tests
4bda924 Update master for stable/wallaby
f3a3a61 Imported Translations from Zanata

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I54d65698a758aca82217f57978b00fa234fc5d50