Proposing Xena RC1 for compute-hyperv
This proposed a RC1 release for compute-hyperv to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 12.0.0..1971a67
d508d59 doc: update IRC information
700e6b4 Drop OVS support
b85e6a2 Bump requirements
fc24385 setup.cfg: Replace dashes with underscores
e0aceaa Add Python3 xena unit tests
8e85d70 Bump minimum requirements to match Nova
12c7fb6 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I681ebe1388e7c89771697aec8c00243ebc4a0cc2