Proposing Xena RC1 for ec2-api
This proposed a RC1 release for ec2-api to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 12.0.0..2a5a973
2a5a973 doc: Update our IRC server to OFTC
084d252 [ussuri][goal] Update contributor documentation
c0f612e setup.cfg: Replace dashes with underscores
8775709 Add Python3 xena unit tests
cdb6eef fix using of SQLAlchemy.in_ operator in UT

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ief9da946585a416743f7031144a720cfc99e6277