Proposing Xena RC1 for freezer-web-ui
This proposed a RC1 release for freezer-web-ui to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 10.0.0..39157c0
abf50fe [ussuri][goal] Update contributor documentation
583ba22 setup.cfg: Replace dashes with underscores
5d82e7a Add Python3 xena unit tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I9e76369983a628e045e75e4ee6decda2dae2f5f1