Proposing Xena RC1 for freezer
This proposed a RC1 release for freezer to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 10.0.0..38cb8cd
38cb8cd Moving IRC network reference to OFTC
a495023 Add lower-constraints gate CI
32337c8 Fix the error of  tempest test cases for schedule.
c33d8d9 Update "if .. if .." to "if .. in .."
ea30ed7 setup.cfg: Replace dashes with underscores

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If79a4d66626224cd33031ec31cb70df7d3e4fd2d