Proposing Xena RC1 for heat-agents
This proposed a RC1 release for heat-agents to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 2.2.0..ba45ab4
70b1d06 [ussuri][goal] Update contributor documentation
7c62d3b Migrate from testr to stestr
b4d249a Add Python3 xena unit tests
6dc9414 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ie96f407b6cdef57555aaca633bbd8fb4c88489d3