Proposing Xena RC1 for kuryr-libnetwork
This proposed a RC1 release for kuryr-libnetwork to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 7.0.0..dbb7a4d
dbb7a4d setup.cfg: Replace dashes with underscores
f2edb0c Add Python3 xena unit tests
03958c9 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic42b9ca62e4e323de835bc6b801f93ef938b08d9