Proposing Xena RC1 for magnum-ui
This proposed a RC1 release for magnum-ui to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 8.0.0..bcd7fbf
bcd7fbf Default `floating_ip_enabled` to true in cluster template
c2a6acc Update horizon nodejs template
565614f Add overlay2 to Docker storage driver dropdown
1fc627f Support zero workers nodes on Horizon dashboard
335fb71 Remove hard_limit check in dashboard
c9555d0 Add Python3 xena unit tests
74a0261 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I9fb404fd2bc2b025c65a50c689769060d62dacea