Proposing Xena RC1 for manila-ui
This proposed a RC1 release for manila-ui to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 5.0.0.0rc1..299e102
3d499a6 Follow up patch for API v2.45, fixes bug on rule metadata form
748b25f API v2.46, added 'is_default' field to share group types
957240e API v2.46, added 'is_default' field to share types
8b5b9cb API v2.45, adds metadata support to access rules
4d33369 Set parent share az when creating a share from snapshot
eb52c16 Changed minversion in tox to 3.18.0
803ab32 Fix share group creation with share types
ba07497 [CI] Temporarily set integration test job to non-voting
37e5b2f Fix parsing names in switched fields
cc3cfad setup.cfg: Replace dashes with underscores
4edb837 Remove nova networking
89a90ff Support policy-in-code and deprecated policy
7cb7f1c Add Python3 xena unit tests
af7122c Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ia7872673bb8af26065cfcf23ca87976fbe01c8a7