Proposing Xena RC1 for masakari-dashboard
This proposed a RC1 release for masakari-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 4.0.0..610116b
610116b [CI] Run bandit
39f0306 Update policy file to yaml
a01dacc [docs] IRC moved to OFTC
f50f432 Change minversion of tox to 3.18.0
934a039 [community goal] Update contributor documentation
b537457 Replace getargspec with getfullargspec
4260038 setup.cfg: Replace dashes with underscores
a7fdb13 Add Python3 xena unit tests
f09f8f2 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic2830c3dde8affa623505ae52e723270c35b78de