Proposing Xena RC1 for mistral-dashboard
This proposed a RC1 release for mistral-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 12.0.0..2980dfc
2980dfc [community goal] Update contributor documentation
fa785b2 setup.cfg: Replace dashes with underscores
a4f8e09 Use py3 as the default runtime for tox
6b73bda Add Python3 xena unit tests
ce7a949 Update master for stable/wallaby
7fd7c82 Setting DEFAULT to False.
f5473fc Fix gate jobs

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ief2a74f01f038f6a862590aa35c060ce1347c143