Proposing Xena RC1 for monasca-api
This proposed a RC1 release for monasca-api to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 6.0.0.0rc1..7466a2f9
7466a2f9 Replace removed DB methods with current ones
3d43cbec setup.cfg: Replace dashes with underscores
5b2af7db Fix Docker build
a6e56ab6 Add Support for Falcon 3.0.0, 3.0.1
ae5dffc9 [Doc] Update Freenode to OFTC as our IRC server
1b5cd302 Fix log-metric configuration
545c7a08 Add Python3 xena unit tests
4f12b214 Update master for stable/wallaby
a93e1cc9 Workaround for new pip 20.3 behavior
64d62a54 Fix warnings in documentation

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ia79fcc48354b5838705b70720ac5f179d20a8eae