Proposing Xena RC1 for monasca-notification
This proposed a RC1 release for monasca-notification to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 4.0.0..fa2454a
fa2454a Fix Grafana URL in email notification
54f66e9 Add Python3 xena unit tests
ff1872d Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ibd0fc4b7ce6aa3109f7d14957fd5010fb22f5dcb