Proposing Xena RC1 for monasca-persister
This proposed a RC1 release for monasca-persister to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 4.0.0..cbc6541
31152e1 Fix Docker build
72c0b69 Fix zuul publish docker image job
dc589ae Add Python3 xena unit tests
5d25622 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I2a6a6b1ae641f0ca2067af55b6e27b79caad56a2