Proposing Xena RC1 for monasca-thresh
This proposed a RC1 release for monasca-thresh to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 4.0.0..d256f05
37ccf62 Revert "Use external Zookeeper in Local mode"
d54904e Fix Docker build
d156980 Use external Zookeeper in Local mode

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I436835a0c7a32bc656fbbbeee6ed333ccdd1393d