Proposing Xena RC1 for monasca-ui
This proposed a RC1 release for monasca-ui to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 4.0.0..9014bce
9014bce Add Python3 xena unit tests
e90fb18 Remove prefix 'dim_' from dimensions in Drilldown Dashboard
36cc4be Add missing PY3 support in overview views

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I555c9a8482e90c5b5ba9e6474153a785e66baf0a