Proposing Xena RC1 for murano-agent
This proposed a RC1 release for murano-agent to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 7.0.0..db57214
db57214 [community goal] Update contributor documentation
89b4512 Fix lower-constraints job
80bbf4e Add Python3 xena unit tests
3323cc5 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Id8c90605f8476a56b0692c972d847e214330712f