Proposing Xena RC1 for murano-dashboard
This proposed a RC1 release for murano-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 11.0.0..fb7d081b
fb7d081b Fix horizon compilemessages with murano ui enabled
17562bb2 Drop horizon-nodejs10-jobs template
21a1209b [community goal] Update contributor documentation
80df3407 setup.cfg: Replace dashes with underscores
6175248c Add horizon-nodejs14-job template.
b88b32ab Add Python3 xena unit tests
7c6eeb71 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I05c919ca22cb5320e6a2de41d9e9d105ac5d62c3