Proposing Xena RC1 for networking-baremetal
This proposed a RC1 release for networking-baremetal to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 4.0.0..cd69016
cd69016 Add lower-constraints job to current development branch
ae68d69 Increase version of hacking and pycodestyle
5bcf940 Update min version of tox to use allowlist
0fb7388 setup.cfg: Replace dashes with underscores
a0bd483 Add Python3 xena unit tests
14eaffe Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic7eaaf2d9457c922080f7b14b1189ae5c216b2ec