Proposing Xena RC1 for networking-generic-switch
This proposed a RC1 release for networking-generic-switch to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 5.0.0..6b8b7b5
6b8b7b5 Add lower-constraints job to current development branch
1ab2f8d Increase version of hacking and pycodestyle
903e969 Fix tempest based job
1828169 Update min version of tox to use allowlist
33f7bda setup.cfg: Replace dashes with underscores
f9900d2 Fix command syntax in Juniper driver
1eebd8f Add Python3 xena unit tests
5741716 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I5ce81c9be68b64355d5d1d421ea2e953d9631f2c