Proposing Xena RC1 for networking-hyperv
This proposed a RC1 release for networking-hyperv to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 10.0.0..2329562
2329562 doc: update IRC information
a913cc6 setup.cfg: Replace dashes with underscores
60a5b5e Add Python3 xena unit tests
bcafa01 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I24bee40b154725ee9e4d1f6314a56428e279dfd9