Proposing Xena RC1 for networking-odl
This proposed a RC1 release for networking-odl to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 18.0.0..40ca37f63
78c5c64c2 Changed minversion in tox to 3.18.0
d01304a67 Use assertCountEqual instead of assertItemsEqual
59dd774b7 Use Payloads for FLOATING_IP
8ba1e9a73 Use payloads for PORT SUBNET and SECURITY_GROUPS events
6b07846c9 Enable tls-proxy support for test job
2470f7f02 Retrieve the agent using "get_agent_objects"
e6f936925 Explicitly set jobs to ML2/OVS
5fd279630 setup.cfg: Replace dashes with underscores
5ae7a90d4 use payloads for TRUNK and SUBPORTS callbacks
732a90818 Change the remaining non python3 compatible code
98cb8439a Update master for stable/wallaby
8803f2592 Add Python3 xena unit tests
9a57c5c04 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Iae039d8c60afa2896893de4e9e28fdd1d875271e