Proposing Xena RC1 for networking-sfc
This proposed a RC1 release for networking-sfc to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 12.0.0..04e1be1
f211d5d [functional] Remove duplicate code from parent test class
21dce30 Use ovs constants from neutron-lib
668031f Enable tls-proxy support for test jobs
e01048c Use assertCountEqual instead of assertItemsEqual
069b36b pdf-docs: use UPPER_CONSTRAINTS_FILE for doc build
ae040dc Add Python3 xena unit tests
0677479 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ife052392198da18a331d4f6c718b650e6ff62af5