Proposing Xena RC1 for neutron-dynamic-routing
This proposed a RC1 release for neutron-dynamic-routing to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 18.0.0..891ca86
891ca86 use payloads for ROUTER_INTERFACE events
b8b8788 Use payload for PORT and FLOATING_IP events
c021b10 Use bgp api-def from neutron-lib
7245c22 setup.cfg: Replace dashes with underscores
f9288d7 Use py3 as the default runtime for tox
43f50d5 Fix TypeError when formatting BGP IP address.
28c8320 Add Python3 xena unit tests
19347ce Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I13f751c4502ca9597571bd38bd3eee0293c90de3