Proposing Xena RC1 for neutron-vpnaas
This proposed a RC1 release for neutron-vpnaas to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 18.0.0..8de26cbfc
8de26cbfc Use payload for ROUTER BEFORE_UPDATE event
bd75769c8 Moving IRC network reference to OFTC
52eba441d Fix functional jobs due to OVS file removal
229709887 Add ipsec.secrets reload function to strongSwan driver
68ac84145 Add Python3 xena unit tests
42ae0626b Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Icbfc237d0c430cbdf65a4bb59484f5bde9c690ea