Proposing Xena RC1 for octavia-dashboard
This proposed a RC1 release for octavia-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 7.0.0..33c308d
7e39031 Change the Octavia Barbican namespace
91042f2 Drop horizon-nodejs10-jobs template
27f5295 Imported Translations from Zanata
cd1f06a Add example logging setting
ff3addc Add horizon-nodejs14-job template.
4421ff8 Fix deleting multiple resources
7fb4bac Add Python3 xena unit tests
17ffab7 Update master for stable/wallaby
f469a22 Remove unicode to adapt to python3

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ie6845bcb915acfaf11bbe990fa61b1138d74b1ce