Proposing Xena RC1 for sahara-plugin-cdh
This proposed a RC1 release for sahara-plugin-cdh to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 5.0.0..fc27bc3
fc27bc3 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I614138b4f1ef96ca098525984e0109dc90a69a25