Proposing Xena RC1 for sahara
This proposed a RC1 release for sahara to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 14.0.0..2ce24b6f
2ce24b6f Remove six
e0136189 Preinstall requirements for docs build
0e200492 Drop support for Block Storage API v2
622af694 docs: Update Freenode to OFTC
a0de19e4 health check error about SqlAlchemy
5d2f114b delete validation error cluster
9a69cc52 [goal] Deprecate the JSON formatted policy file
5e84918d Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ibfabca15847cc478e85984049f90ef5e6e0964e1