Proposing Xena RC1 for senlin-dashboard
This proposed a RC1 release for senlin-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 3.0.0..fef54bb
fef54bb Drop horizon-nodejs10-jobs template
1c4a055 Add horizon-nodejs14-job template.
de1817a Add Python3 xena unit tests
ba63dc7 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I15c810175dcfb595e6b33aef932fedd75e6a7600