Proposing Xena RC1 for senlin
This proposed a RC1 release for senlin to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 11.0.0..2216af78
2216af78 Replace deprecated inspect.getargspec
3e02c0c8 Moving IRC network reference to OFTC
cd33c125 senlin.policy.loadbalance module fix key error
5922e74a Fix LB creation failure with cluster name
412197ad setup.cfg: Replace dashes with underscores
7404f76a add senlin cluster_name for lb/pool/listener name for creation
755421a6 When session_persistence type NONE set option None
3b121536 Remove lower constraints
64658fef Add Python3 xena unit tests
69f1824b Update master for stable/wallaby
b262314b Adding PERSISTENCE_TYPES option NONE
40830c88 Fix health policy attach/detach action when messaging.MessagingTimeout occurred
1bace464 Update error in install-rdo.rst and update default port
7f7f5d65 Correct a typo in the document

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic71fae59105314f1a70fba9d1169460e0b87b164