Proposing Xena RC1 for solum
This proposed a RC1 release for solum to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 10.0.0..51b5559
51b5559 Moving IRC network reference to OFTC
6c3bfa7 Add Python3 xena unit tests
1efe736 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I4086c67ee88ad84ec4531770167a96a892c9d36c