Proposing Xena RC1 for tacker-horizon
This proposed a RC1 release for tacker-horizon to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 3.0.0..69a6eb4
b952869 Add tacker-horizon integration test framework
6e80631 Fix retrieval of VNF details in NFV dashboard
3618f21 Drop block storage v2 API
c4f27e2 Remove broken link from tacker horizon guide
028ca19 setup.cfg: Replace dashes with underscores
9a85cce Update min version of tox
ffcb8c4 Add Python3 xena unit tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I9889d89f66d1b949eee4191620d9c335f6d7614f