Proposing Xena RC1 for trove-dashboard
This proposed a RC1 release for trove-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 16.0.0..4225baf
4225baf Fix the configuration issue
05d15b3 Fix log actions
28e0216 Fix nova client initialization in multi-region cloud
d18e3e8 Handle the exception if _extra_data(instance) is called on a missing instance
3c6e8be Change the configuration group panel name
62cd082 Respect OPENSTACK_ENDPOINT_TYPE
2058a23 Add Python3 xena unit tests
6ec4a13 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I3c1146231fae06971b400ef0f3283bfb62175cc4