Proposing Xena RC1 for watcher-dashboard
This proposed a RC1 release for watcher-dashboard to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 5.0.0..4192182
4192182 Fix broken devstack install
a9fc6a8 Enable tox cover jobs on zuul
5d22d8c Repair coverage report tox job
a48e09f Changed minversion in tox to 3.18.0
8a2d2eb Manage constraints with testenv install_command
0d9323e Add Python3 xena unit tests
d147750 doc: Update our IRC server to OFTC
2ee3c60 Fix create audit fail from audit_template
0aee486 Use py3 as the default runtime for tox
54a6e10 setup.cfg: Replace dashes with underscores
427afe1 Dropping lower constraints testing
399b3cd Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I84a95e636a7e750c2d10f886ed57765576dad63d