Proposing Xena RC1 for watcher
This proposed a RC1 release for watcher to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 6.0.0..6dd2f2a9
6dd2f2a9 BugFix:  Prevent float type variables from being passed to random
a9938499 Use Block Storage API v3 instead of API v2
c28756c4 use HTTPStatus instead of direct code
2414f66e Add watcher dashboard to devstack documentation
546b730c Changed minversion in tox to 3.18.0
e0779175 Manage constraints with testenv install_command
1235153b Use py3 as the default runtime for tox
204992ff Enable tls-proxy as default in test jobs
6f668133 setup.cfg: Replace dashes with underscores
9ca44fa3 Replace deprecated with_lockmode with with_for_update
2205f4e4 Add Python3 xena unit tests
7cd43737 Update master for stable/wallaby
a2123088 requirements: Drop os-testr

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ib5e186a86d8fa24e82c79f4c9743e1e87a37280f