Proposing Xena RC1 for zun
This proposed a RC1 release for zun to get any merged changes
out prior to the RC1 deadline for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 16.

$ git log --oneline --no-merges 7.0.0..abd1e5c9
abd1e5c9 Respect allocation ratio on container update
1f105766 Deprecate Aggregate[CPU|Ram|Disk]Filters
d8b7b01d Replace deprecated with_lockmode with with_for_update
52a35609 setup.cfg: Replace dashes with underscores
917868f5 Add Python3 xena unit tests
ae2aa6a0 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I69128928d9abe5f9ece4b65be2ce4c930ada34b6