Release final python-adjutantclient for Xena
This proposed a release for python-adjutantclient to get any final merged
changes out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 17.

$ git log --oneline --no-merges 0.8.0..20d234b
20d234b Migrate from testr to stestr

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic47da4d95890932c543330a04f11c1047723563c