Release final python-barbicanclient for Xena
This proposed a release for python-barbicanclient to get any final merged
changes out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 17.

$ git log --oneline --no-merges 5.1.0..ad49c40

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I7ba606e6b3a0d3c24982e7bee442518c4d4db5fc