Release final python-freezerclient for Xena
This proposed a release for python-freezerclient to get any final merged
changes out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 17.

$ git log --oneline --no-merges 4.2.0..a5496f7
a35d808 Fix lower-constraints CI error
f10f3ae [ussuri][goal] Update contributor documentation
c50503c Use py3 as the default runtime for tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I04733290d003fef0fbdc227e3c7f16899acb3013