Release final python-masakariclient for Xena
This proposed a release for python-masakariclient to get any final merged
changes out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 17.

$ git log --oneline --no-merges 7.0.0..d5b24b2
e3c1bb1 Fix home_page in setup.cfg
0014e72 Fix url to docs
30eff27 [CI] Run bandit
6d96104 [community goal] Update contributor documentation
d337fbe Replace deprecated UPPER_CONSTRAINTS_FILE variable
75a4c80 setup.cfg: Replace dashes with underscores
dd83634 Use py3 as the default runtime for tox
9768f59 Add Python3 xena unit tests
f615744 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I135919402c7d867ad0ab8b800d7913f7f2113d35