Release final python-muranoclient for Xena
This proposed a release for python-muranoclient to get any final merged
changes out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 17.

$ git log --oneline --no-merges 2.2.0..c3dd963
c3dd963 [community goal] Update contributor documentation
1145c47 Add Python3 xena unit tests
04cc702 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I50ea0873951162c442d21faf24cb722f7fbdfebc