Release final python-solumclient for Xena
This proposed a release for python-solumclient to get any final merged
changes out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 17.

$ git log --oneline --no-merges 3.3.0..e43cee9
7156485 Fix lower-constraints job
2be3c97 Add Python3 xena unit tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I33e5deaf767427036a5ac6f15051efe3c3aa48fb