Release final python-zaqarclient for Xena
This proposed a release for python-zaqarclient to get any final merged
changes out prior to the lib freeze for Xena going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 17.

$ git log --oneline --no-merges 2.1.0..e388947

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ib5e8513ffb1a3ba427f0c1126b8953be467bd519