Release cloudkitty-tempest-plugin for xena
cloudkitty-tempest-plugin is a tempest plugin. This patch propose a latest release for
cloudkitty-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 2.4.0..b1117d5
94121d4 Use py3 as the default runtime for tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I1409185e63f50358743ab960f63655f93ff8a487