Release freezer-tempest-plugin for xena
freezer-tempest-plugin is a tempest plugin. This patch propose a latest release for
freezer-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 1.3.0..de9f456
1c5a831 Enable tls-proxy support for test job

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If3270ea88976207b46a53262102b0a5b2ec3627d