Release heat-tempest-plugin for xena
heat-tempest-plugin is a tempest plugin. This patch propose a latest release for
heat-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 1.3.0..adb8ba7
335c11b Use Block Storage API v3 instead of API v2
fcdf286 Make all functional jobs non-voting
67f98f8 Migrate from testr to stestr

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I654442610499e7d7703ba2486be02ef29137258f