Release ironic-tempest-plugin for xena
ironic-tempest-plugin is a tempest plugin. This patch propose a latest release for
ironic-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 2.3.0..663ced6
7504e63 remove ebtables fix env var from wallaby stable job

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I156bc63e4bb4126e49199606e36c57d3f4e71ecc