Release kuryr-tempest-plugin for xena
kuryr-tempest-plugin is a tempest plugin. This patch propose a latest release for
kuryr-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 0.10.0..19a3314
19a3314 Check LB members when scaling a deployment
02778b3 Check service longevity

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I6c654ffef71fd081499656fd1691d01631f72f39