Release murano-tempest-plugin for xena
murano-tempest-plugin is a tempest plugin. This patch propose a latest release for
murano-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 2.3.0..e3ff450
e3ff450 [community goal] Update contributor documentation

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ica172fe2ccade91315d2e3af859f90800c4f7f84