Release octavia-tempest-plugin for xena
octavia-tempest-plugin is a tempest plugin. This patch propose a latest release for
octavia-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 1.8.0..605899c
605899c Increase job timeout for noop-api tests
6855323 Fix wrong protocol in UDP pool tests
5cf1f61 Fix two-node job configuration
1c53f1b Add stable/wallaby jobs to check and gate queues
88d78fb Fix expect_timeout_error in IPv6 scenario tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I34e64d0095446d75bc2add5e3b8df2b20cacce1d