Release telemetry-tempest-plugin for xena
telemetry-tempest-plugin is a tempest plugin. This patch propose a latest release for
telemetry-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 1.4.0..c410268
d5c98fd Focus testing on train and wallaby
aade825 Make telemetry-dsvm-integration patches vote again
d42dc01 Adding FIPS job definitions for telemetry components
34e514a Remove autoscaling until gate issues sorted out

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Idab41e9703c7a5937f01b94c0576c1c71f19d296