Release zaqar-tempest-plugin for xena
zaqar-tempest-plugin is a tempest plugin. This patch propose a latest release for
zaqar-tempest-plugin to get any merged changes out prior to the final Xena going
into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, September 23.

$ git log --oneline --no-merges 1.3.0..3813c99
3813c99 Migrate from testr to stestr

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I9174c6575b1556fa5152727ba18405f71e04f6d2