Final RC patch for monasca-ui
This creates a new RC before the final release candidate deadline to
release the latest fixes and/or translations.

Please ack with a +1, or if there is some reason to hold off, leave a -1.

Patches with no response will be merged by Thursday, September 30.

$ git log --oneline --no-merges 5.0.0.0rc1..ec5420d
ec5420d Update TOX_CONSTRAINTS_FILE for stable/xena
8160cd1 Bump decorator version in lower-constraints.txt
36cbdd1 Update .gitreview for stable/xena

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I48eadbaa26017ae94a533c6bf74e2b908ea53f95