Xena final releases for cycle-with-rc projects
This tags the final release for all cycle-with-rc projects for Xena.
While not required, having PTLs ack this patch would be nice to include
that record in the merged metadata.

Release activity should now be frozen until this is processed on Oct 6.
Any bugfixes merged to stable/xena can be queued up and ready to go
as a follow on stable release after the coordinated release date.

There are different diff-start values for each project. This is used
when generating release notes to denote the point where the
stable/wallaby branch started to diverge from what is included on
stable/xena. For some, this will be the final Wallaby major version,
but if multiple RCs were done, this may show the initial RC1 release
version. This is normal.

This patch will be merged on Oct 6.

Change-Id: Ic315f2d0b9377975a9cf72fe284f71cb090ace5c