Update EM transition date of Victoria and Xena
This patch propose to move the Extended Maintenance transition date of
stable/victoria with one week to have a less busy period of almost ~1
month after the release of Yoga to prepare.

Also updates the EM transition date of Xena to 18 months after the
initial release. This will most probably change in the future, but
let's have a more valid date there then the current one.

Change-Id: I7fec394cd914967e54b1a4651c2d2ed0ba9def37