[ironic] Switch other projects to CWI
This commit aims to switch the following deliverables to
cycle-with-intermediary:
* ironic-prometheus-exporter
* ironic-ui
* networking-baremetal
* networking-generic-switch

We understand that this deliverables are not released very often
during the cycle, but having them as cycle-with-rc can cause a false
impression that we added/removed things that would cause a major bump
version. All other ironic deliverables are following the
cycle-with-intermediary model. For more context please read [1].

[1] https://meetings.opendev.org/irclogs/%23openstack-release/%23openstack-release.2021-09-27.log.html

Change-Id: I01664d12178f86bbafe763b620d3a4ec06353436