Release horizon 20.2.0(Yoga)
This is the first release for Yoga. The minor version is bumped as we
have changes in requirements.txt(Django>=2.2,<3.0 changed to >=2.2,<3.3).
Horizon bumps the upper version of Django to Django3.2. So this patch
cut a new release of the horizon so that all plugins can consume
this change. For more information please see [1]

[1] https://review.opendev.org/c/openstack/horizon/+/815393

Change-Id: Ie0e7c22fff759016b10b3f4c143928323dfb90b4