Document how to fix release circular dependencies
openstack/releases has a number of dependencies, and some of them (reno,
pbr, openstack-governance) are released using openstack/releases itself.

This may create circular dependencies issues, if one of those
dependencies becomes uninstallable: in that case we won't be able to use
openstack/releases to release a fixed version.

The proper way to solve this issue is to push a tag manually to trigger
the release of the affected repository, then retroactively push the
corresponding openstack/releases change. This will be a noop tag-wise
but make sure that the metadata is correct going forward.

This patch leaves comments in the affected files, so that we remember
how to solve this next time it happens!

Change-Id: I695e83410ba704e9978147e226e9555890a848a8