[OpenStackAnsible Roles] Transition Pike to End of Life
This transition the Pike branch to End of Life. The last patch of the
branch will be tagged with ocata-eol tag. The pike branch cannot be
used anymore and will be deleted if this patch merges.

This is needed as stable/pike is not actively maintained in the recent
period and thus gate is broken due to job failures. By removing these
branches, infra resources will be freed up, too.

Change-Id: Iaf4fdd8bf43c13c73ea7b2b35b50d457a9fada37