[OpenStackAnsible Roles] Transition Queens to End of Life
This transition the Queens branch to End of Life. The last patch of the
branch will be tagged with queens-eol tag. The queens branch cannot be
used anymore and will be deleted if this patch merges.

This is needed as stable/queens is not actively maintained in the recent
period and thus gate is broken due to job failures. By removing these
branches, infra resources will be freed up, too.

Change-Id: I589a603bd05587130ff9f70cc91c3090817d6ac0