Yoga-2 Release for oslo.service
This is the Yoga milestone-2 release for oslo.service.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

$ git log --oneline --no-merges 2.7.0..6552b9a
6552b9a Make debug option of wsgi server configurable

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I0d1110ef372d24155977a2a3262ec3ea8098dccd