Yoga-2 Release for python-aodhclient
This is the Yoga milestone-2 release for python-aodhclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

$ git log --oneline --no-merges 2.3.0..3a36ed1
3a36ed1 Fix aodhclient for pyparse 3.0.6

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ib3320df0c350918d9d53c7eb0f6bddf13bc3e878