Yoga-2 Release for python-heatclient
This is the Yoga milestone-2 release for python-heatclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

$ git log --oneline --no-merges 2.4.0..db9a71f
db9a71f Update python testing classifier
796a8f5 Replace deprecated import of ABCs from collections
4753a47 Fix test and lower constraints
b457930 Add Python3 yoga unit tests
3415690 Update master for stable/xena

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I1d4c5d82ba464f0a917c075471101b1fec378020