Yoga-2 Release for python-ironicclient
This is the Yoga milestone-2 release for python-ironicclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

$ git log --oneline --no-merges 4.9.0..4549820
4549820 Improve description of "node boot mode set"
4ce3c4e Test python 3.6 for distributions compatibility
4712812 Made `baremetal <command> --help` display command specific help.
e76c43b Fix references to ironicclient classes and methods

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ia25e1971e884cc34738cb89ef4f4c232124382f2