Yoga-2 Release for python-keystoneclient
This is the Yoga milestone-2 release for python-keystoneclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

$ git log --oneline --no-merges 4.3.0..100253d
56c7b50 Add access to /v3/auth/systems
f6569e2 Fix doc error to unblock the gate
0c4e294 Drop lower-constrait job

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic5c870f567a724f23d806e210dd557c66ba5d12e