Yoga-2 Release for python-tackerclient
This is the Yoga milestone-2 release for python-tackerclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

$ git log --oneline --no-merges 1.8.0..b7f27c3
b7f27c3 Update python testing classifier
6834454 Support of Cancel VNF command in openstackclient
a3d485c Add Python3 yoga unit tests
861a575 Drop test for lower constraints
015e6e2 Update master for stable/xena
b6aec20 setup.cfg: Replace dashes with underscores
4299140 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I39794072d0628cc4fa64272dc80860dbddd23846