Yoga-2 Release for python-zunclient
This is the Yoga milestone-2 release for python-zunclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

$ git log --oneline --no-merges 4.3.0..2284443
2284443 Python 3.10: Subcomand help test fix
24edb14 Add Python3 yoga unit tests
faf7563 Update master for stable/xena
dd73280 Remove the unused os-testr package

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I41d5e7fd26ecf232017bebfc11f0a4e929e9b580