Yoga-2 Release for oslo.cache
This is the Yoga milestone-2 release for oslo.cache.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 06.

Change-Id: Ia3f949eb7a5aec945af57fe950b15f32849d6a13