Release bifrost for stable/victoria
This release picks up new commits to bifrost since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 9.0.1..4d1b0cae
4d1b0cae [CI] Move Kolla testing to CentOS Stream
1487806e Revert "libvirt: pin edk2-ovmf for CentOS Stream"
15bcd495 Recognize TOX_CONSTRAINTS_FILE as a valid constraints variable
7912fc5a Install DIB dependencies in bifrost-ironic-install role
822b65a6 Add OFTC network channel to contributor docs
93f378f5 libvirt: pin edk2-ovmf for CentOS Stream
f2e6bb82 Point ipa-builder to stable/wallaby
a2b16602 Install squashfs-tools for Ubuntu image builds
c2a4f845 Use explicit path for which
8f11a86a Increase memory of vm to adjust for new tinyipa
d4f60272 Upgrade pip version to fix pep-517 enforcing
22efef9a Fix the Debian Buster version number
1d84631d Restore the conditional verbosity in bash scripts
d8fec0a0 Misplaced variable preventing fast-track after inspection
7c42c1e3 Exit on all errors in bash scripts when possible
1855fbdb Install at least pip version 19.1.1
e6fde210 Collect firewalld info if present
cb3e8104 docs: explain OS support in terms of two tiers
0cd0c1de Fix passing parameters with spaces to bifrost-cli
daddfa40 Remove lower-constraints job
c983acfa Switch to CentOS 8 Stream for testing
4eb06881 Fix two CI issues
102979ca Fix missing node facts with BIFROST_INVENTORY_SOURCE=ironic

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I367968f47238ae77039490230ade92055039c868