Release blazar for stable/victoria
This release picks up new commits to blazar since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.0.0..6fc1d3c
6fc1d3c docs: Update Freenode to OFTC
bfba274 Fix use of legacy Alembic API
e0bd740 Remove lower-constraints job on stable branches
002024d Update TOX_CONSTRAINTS_FILE for stable/victoria
253d10a Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I18e9743cb9bec82be9e8870c7b3b3d0cfe6c1c52