Release ceilometer for stable/victoria
This release picks up new commits to ceilometer since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 15.0.0..e33a1f7d
e33a1f7d Remove jobs corresponds to obselete featuresets
3f73d8ba Fix CA file for Swift pollster
4bb02413 Do not install libvirt python bindings from pip
564a2ca6 [Stable-only] Pin Gnocchi to a specific release
418af55c [stable-only] Add font for PDF generation and fix lower constraints
e953512d Cap tenacity < 7.0.0
e555e938 Update TOX_CONSTRAINTS_FILE for stable/victoria
bb6dda05 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I0dd2a898d7b47f118537c31e52ed1ac252cefdc8