Release cloudkitty-dashboard for stable/victoria
This release picks up new commits to cloudkitty-dashboard since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.0.0..2a24f55
2a24f55 Respect endpoint type specified in Horizon
ea6adc2 Update TOX_CONSTRAINTS_FILE for stable/victoria
fbb4510 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ib6107c3f900b7946d34d2e0c39bcdf8bb6d0d50c