Release ec2-api for stable/victoria
This release picks up new commits to ec2-api since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.0.0..2fff77e
b2686a2 various fixes
ee4f8f9 Run TripleO jobs on CentOS8 instead of CentOS7
4fec7cd Update TOX_CONSTRAINTS_FILE for stable/victoria
7102fec Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I1d016ce777798891367aa1f772ad7c00ced6272e