Release heat-dashboard for stable/victoria
This release picks up new commits to heat-dashboard since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..02d04ec
02d04ec Imported Translations from Zanata
080771f Run npm nodejs job with Firefox browser
e601aab Move image_field_data method in-tree
85d2a88 Imported Translations from Zanata
44765bb Update TOX_CONSTRAINTS_FILE for stable/victoria
36565ac Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I7efe2b417d3203586cbd7b6c3e3b27c2dd7f7b73