Release heat for stable/victoria
This release picks up new commits to heat since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 15.0.0..cbe9be5e7
cbe9be5e7 Provide install_command to pip install with upper-constraints
c062de4f2 Fallback to upstream for Fedora image
f52feef69 Revert "Make grenade job voting again"
7ef31c2be Make grenade job voting again
86354c2d7 Don't create deployment when changing transport
40a3494c8 Update get_hosts to use available API
c881da8b8 Use Block Storage API v3 instead of API v2
98fade887 Detect EL8 platform-python
0833daa3f Preserve order in list_concat_unique
652f872e4 Resume testing grenade jobs in Victoria
59328b254 Assume interface detach called for port not found
8bdf6b984 Don't update status for replaced resource
546bc2391 Net attr: Sort segments without name attribute first
d8027e65c Fix args for initializing trove client
915da60e8 Fix multiple gate issues
7e03502a6 Align lower-constraints for new pip
779b39c97 Replace Fedora test image with F32
3cf3a68a4 Add SOURCE_IP_PORT to LB_ALGORITHM allowed values
996884da6 Delete the default value of domain in role creation
3f98fd9b1 As of nova microversion 2.57, personality is deprecated
e30ea0af5 Configure swap as before
5f1ab43bb Update TOX_CONSTRAINTS_FILE for stable/victoria
dc6602079 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I0070cff25fd6f77ea488dc5837e2d769cb03618e