Release ironic-lib for stable/victoria
This release picks up new commits to ironic-lib since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.4.1..b117e63
b117e63 utils.execute: log stdout and stderr even on failure
f4cabc0 Limit the number of malloc arenas for qemu-img convert
d27b4a2 Raise qemu-img memory limit to 2 GiB

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ic2859b8be6cb19d380948a11e4872c69b035c09c