Release ironic-python-agent for stable/victoria
This release picks up new commits to ironic-python-agent since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.4.3..0b292a1f
0b292a1f Re-read the partition table with partx -a, part 2
06cf7f3d Re-read the partition table with partx -a
9525a115 Fix UEFI record regex
47ac40a7 Delete EFI boot entry duplicate labels first
750934a2 Output verbose info from efibootmgr
48ffbaa0 Force immediate NTP time sync with chronyd at IPA startup
97ce08d0 Fix getting memory size in some lshw output
2645240b Reduce logging verbosity when collecting logs
474d2a91 Utilize CSV file for EFI loader selection
350a67e2 Make _get_efi_bootloaders return relative paths
ee811842 Limit qemu-img execution arenas
601b8330 Fix NVMe Partition image on UEFI
b307c34c Point ipa-builder to stable/wallaby
6e0d08c8 Software RAID: RAID the ESPs
37c39b1b Fix missing data in log messages
bc13adfa Add a call to "udevadm settle" in write_image.sh
b3f489ca Do not fail network interface collection on unsupported interface
92a7d7a9 Fix root UUID for streamed partition images
2076df91 Add fstab pointer to EFI partition
1a19f8b3 Prepare to use tinycore 12
99b9953f Check the base device if the read-only file cannot be read
de1e7512 Increase the memory limit for qemu-img
d1a365e6 Fixes local boot for partition images
0ff57260 Fix error message with UEFI-incompatible images
a820851f Mock tests to return bios boot mode
9961ed62 Prevent broken partition image UEFI deploys
07cf2c1b Option to enable bootloader config failure bypass

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ida8931db2a8727075ead9f43c614532274d9ee3d