Release kayobe for stable/victoria
This release picks up new commits to kayobe since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 9.2.0..a91434f5
e2605a2f Pin rich to <11 due to breaking ansible-lint in kayobe-tox-molecule
07fd6259 Build neutron-sriov-agent image when enabled
0cd50e8c [CI] Stop testing non-stream CentOS
5e444b8a CI: Use correct TD agent repository version
be1d6adb Document that extra kernel parameters are important for inspection
d4717fe1 Add support for Ironic inspection through DHCP-relay
c048c257 Limit ip-routing and snat to seed hosts only

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I021bd7bfb84e57f3538307cfe754cbfc67a7b5bf