Release kolla for stable/victoria
This release picks up new commits to kolla since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.2.0..1c455b5b6
bfc4c8e3c [CI] Stop testing non-stream CentOS
464d15175 Fix variable name

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I5632ab4d3679b64bd52b1abc25b2e0b08cc35f80