Release kuryr-kubernetes for stable/victoria
This release picks up new commits to kuryr-kubernetes since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.0..3b739786
3b739786 Ensure members are deleted from pools when there is no endpoints
9a0344df Add Octavia Tempest Plugin
29043082 Fix base docker images
42154ef7 Fix CI issues
e5180d13 Change registry for fetching coredns image.
c56a2b17 Move to quay.io with kuryr/demo docker image.
d487f227 Update RDO to use Victoria release
2e99333d Increase tempest test timeout to 30 mintues
b7c7bb18 Handle None or {} labels in match_selector()
3e5b0f21 Set read timeout for any request in K8sClient
e7d99389 Update TOX_CONSTRAINTS_FILE for stable/victoria
5be80d4a Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I0ee46427d1db14871961e2b544d2f7e10c763143