Release kuryr for stable/victoria
This release picks up new commits to kuryr since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.1.0..0af5226
0af5226 Remove run_as_root which fails without root_helper
e43bf23 Remove the unused coding style modules
4081dfd Fix pep8 and lower-constraints gates
a338724 Update TOX_CONSTRAINTS_FILE for stable/victoria
35eafe2 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ie37652933b044e0325f309f4cd2aef744b7594a2