Release magnum for stable/victoria
This release picks up new commits to magnum since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.2.0..4b1ba075
4b1ba075 Refix --registry-enabled

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I72c88086255658a35c019bbf81fe57a7a80eb5ea