Release monasca-agent for stable/victoria
This release picks up new commits to monasca-agent since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..91db19f
91db19f Fix Monasca-Setup Keystone detection
4fb0ae6 Fix PY3 issue
574a067 Fix Docker builds
a2d412f Set 'libvirt_type' and 'libvirt_uri' options in the 'libvirt' plugin to be defined in conf.d/libvirt.yaml
1aa5a17 Fix small issue in dokcer plugin.
e340d36 Remove incorrect plugins configuration for docker image.
9db15dd Fix zuul publish docker image job
8be2a8d Update TOX_CONSTRAINTS_FILE for stable/victoria
73d8b18 Update .gitreview for stable/victoria
c0b8ad5 Adding python3-libvirt for Ubuntu-focal

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I6265e114a5b1fc4cd35d120690e53ac32ccc49f0