Release murano-agent for stable/victoria
This release picks up new commits to murano-agent since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.0.0..737fb34
737fb34 Fix base64 usage for Python 3
47116de Update TOX_CONSTRAINTS_FILE for stable/victoria
98642c3 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I1e22ab8d385035bc40c671ec6c11e9906ff14ce8