Release networking-generic-switch for stable/victoria
This release picks up new commits to networking-generic-switch since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..8c1a1c7
8c1a1c7 Fix command syntax in Juniper driver
79aa400 Remove lower-constraints job
f7e74b0 Add documentation for Cumulus Linux NCLU
ef45de9 Update TOX_CONSTRAINTS_FILE for stable/victoria
b4f8536 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I386d00daf4719089deea69fa8cf1ceafae256062