Release octavia-dashboard for stable/victoria
This release picks up new commits to octavia-dashboard since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.0.0..2f1d320
3aef1f7 Change the Octavia Barbican namespace
a66eb7e Imported Translations from Zanata
3954cda Imported Translations from Zanata
a5371c5 Disable Load Balancers panel when Octavia service is not deployed
6f2ef77 Drop lower-constraints job
1b8a97d Imported Translations from Zanata
9737576 Imported Translations from Zanata

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I1974787bc7ca04933fc0333f41a481941f1d32de