Release os-ken for stable/victoria
This release picks up new commits to os-ken since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.2.0..5e9a961c
5e9a961c Avoid missing key 'password' for neighbor_add
e6f2c0d1 Avoid logging MD5 password for BGP add neighbor
c58c8019 Add requirements.txt to tox environment
49976897 Update TOX_CONSTRAINTS_FILE for stable/victoria
df38cf4a Update .gitreview for stable/victoria
b17f6fcb Fix lower-constraints and pep8 for Focal

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I1c41c5d4033cec62169aa1efe39af5049ca857bb