Release oslo.service for stable/victoria
This release picks up new commits to oslo.service since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.4.0..615cfad
615cfad Make debug option of wsgi server configurable
ee19ff3 Unbreak lower constraints
4017909 Update TOX_CONSTRAINTS_FILE for stable/victoria
bb2a341 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Id7665ba8e8904201f322a1102236fb24a55df8db