Release ovn-octavia-provider for stable/victoria
This release picks up new commits to ovn-octavia-provider since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 0.4.0..62891fa
62891fa Ensure that load balancer is added to logical switch
f11eb74 Add log for request_handler events

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I89288a464d9f937fe60631be07e805707c4d4e51