Release ovsdbapp for stable/victoria
This release picks up new commits to ovsdbapp since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.6.0..8c3896a
8c3896a Actually close the connection in Connection.stop()
c2617f9 Don't spam retries 100s of times a second
d75a4df Add an active wait in the "Backend.lookup"
d7cc1b2 Fix docs job
e6bd200 Dropping lower constraints testing (stable Victoria)
4807809 Don't give up when an Exception happens in idl.run

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I48dadbeb0c1e18562042701c50ed5bf0caead2f6