Release python-ironicclient for stable/victoria
This release picks up new commits to python-ironicclient since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.4.0..2e6778f
2e6778f Make baremetal --debug actually work
0187ffb Fix --fields network_data
8ae1c79 Remove lower-constraints job

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ie1fa3096e84253d76ce20c979821c2f7a31bd400