Release python-manilaclient for stable/victoria
This release picks up new commits to python-manilaclient since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.3.0..ab673a2
ab673a2 Fix the id attr for share group type access repr
3feef0c [CI] Fix CI jobs
d1c3a53 Update TOX_CONSTRAINTS_FILE for stable/victoria
394ce6c Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I690adeb5d0074c02ecf3b1159e72ae38e05dd673