Release python-troveclient for stable/victoria
This release picks up new commits to python-troveclient since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.1.1..41a0555
41a0555 Do not make extra API call for listing instances

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: If861dbbe5c411b31e40598f3abe8ea1657d1cf8c