Release sushy for stable/victoria
This release picks up new commits to sushy since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.4.4..ef9bc98
ef9bc98 Revert "Raise an AccessError with SessionService init"
b7a71f1 Raise an AccessError with SessionService init
dba7d1a Fix Processor.sub_processors
efd47e3 Protect Connector against empty auth object

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ib27d51485700555369a9bf51bafbe3f1da9a5b72