Release tacker for stable/victoria
This release picks up new commits to tacker since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..66fd77ab
66fd77ab Drop test for lower constraints
308ed21a Separate nodes of FT and split FT
deb6c2d4 Fix failures and add Update/Scale FT
91cc4c2d Fix a failure in tox lower-constraints
e685613b Turn on FT voting for stable/victoria
4fa0bb88 Temporary deactivate lower-constraints check

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ia96f5d7437aa3ada91a12e9f91c5a727b063789c