Release trove-dashboard for stable/victoria
This release picks up new commits to trove-dashboard since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 15.0.0..1e9ee5b
1e9ee5b Fix log actions
e89a0ab Fix nova client initialization in multi-region cloud
d99d394 Update TOX_CONSTRAINTS_FILE for stable/victoria
41650f6 Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I013180c89dad72f52d69c116c6b24ad8a2b44935