Release watcher-dashboard for stable/victoria
This release picks up new commits to watcher-dashboard since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.0.0..4498a25
4498a25 Fix broken devstack install
745cada Fix create audit fail from audit_template
87e4366 Dropping lower constraints testing
ced2f04 Update TOX_CONSTRAINTS_FILE for stable/victoria
b6942bb Update .gitreview for stable/victoria

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I73124800edd56495bec31749c6e8716377281ce5