Release glance-store for stable/victoria
This release picks up new commits to glance-store since
the last release from stable/victoria.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.3.0..e43a56f
e43a56f Pass multipath config while creating connector object
ff8df6d Drop lower-constraints job
177b803 Update TOX_CONSTRAINTS_FILE for stable/victoria
7f88a4d Update .gitreview for stable/victoria

Change-Id: I2c354abb8b13d31cccbc0129291c5a25df2909dc