Release final blazar-nova for Yoga
This proposed a release for blazar-nova to get any final merged changes
out prior to the lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 18.

Signed-off-by: Daniel Bengtsson <dbengt@redhat.com>
Change-Id: I6ddc46cdfa75b877de4532177408120a84a8f774