[wallaby] release oslo.
Some deliverables never released the patches related to
the `TOX_CONSTRAINTS_FILE` switch for Wallaby. Not landing
this switch downstream (to distros) became problematic as
tox target executed downstream refered to master's
requirements constraints and so validations started to fail
with downstream usage of Wallaby (OSP17 by example).

So even if deliverables does not provides functional changes
I prefer release them as a bugfix version to fix this kind
of problems for distros.

NOTE: oslo.cache is already at the version 2.8.0 on Xena
so I can't release a minor version even if functional changes
are landed.

Change-Id: I920ee7bcca3cbf3e4051b3a3098beed53b1c969d