Release final osc-placement for Yoga
This is a release for osc-placement to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 3.1.1..73e496f
73e496f Add Python3 yoga unit tests
0c24adb Updating python testing as per Yoga testing runtime

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Icf8fa3ed6085f7a524305d24273b361b0f8d0633