Release final python-adjutantclient for Yoga
This is a release for python-adjutantclient to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 0.9.0..20d234b

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ic812bd3f3fb7d94f34ce1734ebdd46bf95ff3e42