Release final python-aodhclient for Yoga
This is a release for python-aodhclient to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 2.4.0..c1d7fe1
c3c0000 Add Python 3 only classifier
20f4372 setup.cfg: Replace dashes by underscores
d978eb7 Add Python3 yoga unit tests
bc7a483 Update master for stable/xena
60df9b4 Add Python3 xena unit tests

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I49686c62547461bb1bc480dadb1a261e98eecf42