Release final python-barbicanclient for Yoga
This is a release for python-barbicanclient to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 5.2.0..ad49c40

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I659b02729dcb00da8b924b59113778b7a0fd33db