Release final python-blazarclient for Yoga
This is a release for python-blazarclient to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 3.3.1..59fd6c0
cf2de7c Add commands for resource property API
9354e62 Updating python testing classifier as per Yoga testing runtime
2091bd9 Add Python3 yoga unit tests
55395d0 Update master for stable/xena

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I0c5d5a890c1040b3e599fd2b7d4e4921925128bf