Release final python-cloudkittyclient for Yoga
This is a release for python-cloudkittyclient to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 4.4.0..783cc22
783cc22 Add response_format to list of available args for v2 summary API
20003a5 Re-add python 3.6/3.7 in classifier
719f4cf Updating python testing as per Yoga testing runtime

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: Ia6d924c49750d667a1e6633ef6fd3864acf94ffd