Release final python-cyborgclient for Yoga
This is a release for python-cyborgclient to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 1.6.0..37026cc
7d563bb bump openstacksdk >=0.46.0
9c26ba7 Add support project-id for arq patch releasenote
63d8d1f Add support for show device profile by name releasenote
2680c35 Refactor the columns generate for device profile list
9d00b7d Updating python testing classifier as per Yoga testing runtime
e1cf238 Get device profile by name in cyborgclient

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I0d29855785f747a5a573330ece4099edec32ed59