Release final python-designateclient for Yoga
This is a release for python-designateclient to get any merged but unreleased
changes out prior to the client lib freeze for Yoga going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, February 25th.

$ git log --oneline --no-merges 4.4.0..485cbdd
52e9ef4 Add profiling support to designateclient
2e26a07 Updating python testing classifier as per Yoga testing runtime
87e9b33 Add basic osc unit tests
15a0070 Add Designate hacking checks to the client
94e7e97 Remove edit-managed from unsupported commands
fa615f7 Use py3 as the default runtime for tox
a3c21fd Cap jsonschema 3.2.0 as the minimal version
d6c78e2 Remove H402 hacking as it is deprecated
f02f009 Stop to use the __future__ module.
5fbc5b1 Clarify that name or ID can be used on TLDs
6a8f10a setup.cfg: Replace dashes with underscores

Signed-off-by: Elod Illes <elod.illes@est.tech>
Change-Id: I4bb03dcc41c3a9deb9a59843a7cc58731528084a